IMPORTANT: Forum
Hey! I added forum.ts for the forum file, so we don't have to mix the codes for user and for forum.

Once you pull the new info, try to check if there are some changes. Might be a good idea to do npm run dev. If it didn't work, write the npm run build and then npm run dev. Thanks!